Split out requestTheme
