+++ client/state/themes/actions/index.js
@@ -2,7 +2,7 @@
 /**
  * External dependencies
  */
-import { map, delay, endsWith } from 'lodash';
+import { delay, endsWith } from 'lodash';
 import debugFactory from 'debug';
 import page from 'page';
 
@@ -11,7 +11,6 @@ import page from 'page';
  */
 import { isExternal } from 'lib/url';
 import wpcom from 'lib/wp';
-import { fetchThemeInformation as fetchWporgThemeInformation } from 'lib/wporg';
 import {
 	ACTIVE_THEME_REQUEST,
 	ACTIVE_THEME_REQUEST_SUCCESS,
@@ -32,9 +31,6 @@ import {
 	THEME_INSTALL,
 	THEME_INSTALL_SUCCESS,
 	THEME_INSTALL_FAILURE,
-	THEME_REQUEST,
-	THEME_REQUEST_SUCCESS,
-	THEME_REQUEST_FAILURE,
 	THEME_SHOW_AUTO_LOADING_HOMEPAGE_WARNING,
 	THEME_TRANSFER_INITIATE_FAILURE,
 	THEME_TRANSFER_INITIATE_PROGRESS,
@@ -62,12 +58,7 @@ import {
 	hasAutoLoadingHomepageModalAccepted,
 	prependThemeFilterKeys,
 } from 'state/themes/selectors';
-import {
-	getThemeIdFromStylesheet,
-	normalizeJetpackTheme,
-	normalizeWpcomTheme,
-	normalizeWporgTheme,
-} from 'state/themes/utils';
+import { getThemeIdFromStylesheet } from 'state/themes/utils';
 import { getSiteTitle, isJetpackSite } from 'state/sites/selectors';
 import isSiteAutomatedTransfer from 'state/selectors/is-site-automated-transfer';
 import { requestSitePosts } from 'state/posts/actions';
@@ -78,98 +69,16 @@ import 'state/data-layer/wpcom/theme-filters';
 
 import 'state/themes/init';
 
-import { receiveThemes } from 'state/themes/actions/receive-themes';
 import { receiveTheme } from 'state/themes/actions/receive-theme';
-import { themeRequestFailure } from 'state/themes/actions/theme-request-failure';
 
 export { setBackPath } from 'state/themes/actions/set-back-path';
 export { receiveThemes } from 'state/themes/actions/receive-themes';
 export { receiveTheme } from 'state/themes/actions/receive-theme';
 export { requestThemes } from 'state/themes/actions/request-themes';
 export { themeRequestFailure } from 'state/themes/actions/theme-request-failure';
+export { requestTheme } from 'state/themes/actions/request-theme';
 
 const debug = debugFactory( 'calypso:themes:actions' );
-
-/**
- * Triggers a network request to fetch a specific theme from a site.
- *
- * @param  {string}   themeId Theme ID
- * @param  {number}   siteId  Site ID
- * @returns {Function}         Action thunk
- */
-export function requestTheme( themeId, siteId ) {
-	return dispatch => {
-		dispatch( {
-			type: THEME_REQUEST,
-			siteId,
-			themeId,
-		} );
-
-		if ( siteId === 'wporg' ) {
-			return fetchWporgThemeInformation( themeId )
-				.then( theme => {
-					// Apparently, the WP.org REST API endpoint doesn't 404 but instead returns false
-					// if a theme can't be found.
-					if ( ! theme ) {
-						throw 'Theme not found'; // Will be caught by .catch() below
-					}
-					dispatch( receiveTheme( normalizeWporgTheme( theme ), siteId ) );
-					dispatch( {
-						type: THEME_REQUEST_SUCCESS,
-						siteId,
-						themeId,
-					} );
-				} )
-				.catch( error => {
-					dispatch( {
-						type: THEME_REQUEST_FAILURE,
-						siteId,
-						themeId,
-						error,
-					} );
-				} );
-		}
-
-		if ( siteId === 'wpcom' ) {
-			return wpcom
-				.undocumented()
-				.themeDetails( themeId )
-				.then( theme => {
-					dispatch( receiveTheme( normalizeWpcomTheme( theme ), siteId ) );
-					dispatch( {
-						type: THEME_REQUEST_SUCCESS,
-						siteId,
-						themeId,
-					} );
-				} )
-				.catch( error => {
-					dispatch( {
-						type: THEME_REQUEST_FAILURE,
-						siteId,
-						themeId,
-						error,
-					} );
-				} );
-		}
-
-		// See comment next to lib/wpcom-undocumented/lib/undocumented#jetpackThemeDetails() why we can't
-		// the regular themeDetails() method for Jetpack sites yet.
-		return wpcom
-			.undocumented()
-			.jetpackThemeDetails( themeId, siteId )
-			.then( ( { themes } ) => {
-				dispatch( receiveThemes( map( themes, normalizeJetpackTheme ), siteId ) );
-				dispatch( {
-					type: THEME_REQUEST_SUCCESS,
-					siteId,
-					themeId,
-				} );
-			} )
-			.catch( error => {
-				dispatch( themeRequestFailure( siteId, themeId, error ) );
-			} );
-	};
-}
 
 /**
  * This action queries wpcom endpoint for active theme for site.
